Renaming the maven module
