--- jms/org.wso2.carbon.transport.jms.jndi/src/main/java/org/wso2/carbon/transport/jms/jndi/provider/JMSServerConnectorProvider.java
@@ -1,48 +1,0 @@
-/*
- *  Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *  WSO2 Inc. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.carbon.transport.jms.jndi.provider;
-
-import org.wso2.carbon.messaging.ServerConnector;
-import org.wso2.carbon.messaging.ServerConnectorProvider;
-import org.wso2.carbon.transport.jms.jndi.listener.JMSServerConnector;
-import org.wso2.carbon.transport.jms.jndi.utils.JMSConstants;
-import java.util.List;
-
-/**
- * Server connector provider for jms.
- */
-public class JMSServerConnectorProvider extends ServerConnectorProvider {
-
-    public JMSServerConnectorProvider() {
-        super(JMSConstants.PROTOCOL_JMS);
-    }
-    public JMSServerConnectorProvider(String protocol) {
-        super(protocol);
-    }
-
-    @Override
-    public List<ServerConnector> initializeConnectors() {
-        return null;
-    }
-
-    @Override
-    public ServerConnector createConnector(String s) {
-        return new JMSServerConnector(s);
-    }
-}
