--- jms/org.wso2.carbon.transport.jms.jndi/src/main/java/org/wso2/carbon/transport/jms/jndi/utils/SerializableCarbonMessage.java
@@ -1,59 +1,0 @@
-/*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.wso2.carbon.transport.jms.jndi.utils;
-
-import org.wso2.carbon.messaging.CarbonMessage;
-
-import java.io.Serializable;
-import java.util.HashMap;
-
-/**
- * Serializable message used to create JMS object message.
- */
-
-public class SerializableCarbonMessage extends CarbonMessage implements Serializable {
-
-    private static final long serialVersionUID = 1;
-    private String payload;
-    private String payloadType;
-    private HashMap<String, String> headers;
-
-    public HashMap<String, String> getHeadersMap() {
-        return headers;
-    }
-
-    public void setHeaders(HashMap<String, String> headers) {
-        this.headers = headers;
-    }
-
-    public String getPayload() {
-        return payload;
-    }
-
-    public void setPayload(String payload) {
-        this.payload = payload;
-    }
-
-    public String getPayloadType() {
-        return payloadType;
-    }
-
-    public void setPayloadType(String payloadType) {
-        this.payloadType = payloadType;
-    }
-}
