--- jms/org.wso2.carbon.transport.jms.jndi/src/main/resources/META-INF/services/org.wso2.carbon.messaging.ClientConnector
@@ -1,1 +1,0 @@
-org.wso2.carbon.transport.jms.jndi.sender.JMSClientConnector
