--- jms/org.wso2.carbon.transport.jms.jndi/src/main/resources/META-INF/services/org.wso2.carbon.messaging.ServerConnector
@@ -1,1 +1,0 @@
-org.wso2.carbon.transport.jms.jndi.listener.JMSServerConnector
