+++ jms/org.wso2.carbon.transport.jms/src/main/resources/META-INF/services/org.wso2.carbon.messaging.ClientConnector
@@ -1,0 +1,1 @@
+org.wso2.carbon.transport.jms.sender.JMSClientConnector
